Extract nested data classes to separate files
- Extract CalculationResult and XirrCalculationResult from CalculationService
- Extract InstrumentEnrichmentContext from InstrumentService
- Extract HoldingKey and HoldingValue from EtfBreakdownService
- Extract HoldingsAccumulator from InvestmentMetricsService
- Extract ValidatedRowData and BasicRowData from LightyearHoldingsService
- Remove temporaryExclusions from ArchitectureTest

Closes #968
