+++ src/main/kotlin/ee/tenman/portfolio/lightyear/BasicRowData.kt
@@ -1,0 +1,6 @@
+package ee.tenman.portfolio.lightyear
+
+data class BasicRowData(
+  val nameParts: List<String>,
+  val weightText: String,
+)
