+++ src/main/kotlin/ee/tenman/portfolio/lightyear/ValidatedRowData.kt
@@ -1,0 +1,8 @@
+package ee.tenman.portfolio.lightyear
+
+data class ValidatedRowData(
+  val name: String,
+  val ticker: String?,
+  val sector: String?,
+  val weightText: String,
+)
