+++ src/main/kotlin/ee/tenman/portfolio/service/HoldingKey.kt
@@ -1,0 +1,7 @@
+package ee.tenman.portfolio.service
+
+data class HoldingKey(
+  val ticker: String?,
+  val name: String,
+  val sector: String?,
+)
