+++ src/main/kotlin/ee/tenman/portfolio/service/InvestmentMetricsService.kt
@@ -44,25 +44,6 @@ class InvestmentMetricsService(
     return quantity to averageCost
   }
 
-  private data class HoldingsAccumulator(
-    val quantity: BigDecimal = BigDecimal.ZERO,
-    val totalCost: BigDecimal = BigDecimal.ZERO,
-  ) {
-    fun applyBuy(tx: PortfolioTransaction): HoldingsAccumulator {
-      val cost = tx.price.multiply(tx.quantity).add(tx.commission)
-      return copy(quantity = quantity.add(tx.quantity), totalCost = totalCost.add(cost))
-    }
-
-    fun applySell(tx: PortfolioTransaction): HoldingsAccumulator {
-      if (quantity <= BigDecimal.ZERO) return this
-      val sellRatio = tx.quantity.divide(quantity, 10, RoundingMode.HALF_UP)
-      return copy(
-        quantity = quantity.subtract(tx.quantity),
-        totalCost = totalCost.multiply(BigDecimal.ONE.subtract(sellRatio)),
-      )
-    }
-  }
-
   fun calculateCurrentValue(
     holdings: BigDecimal,
     currentPrice: BigDecimal,
