+++ src/main/kotlin/ee/tenman/portfolio/service/XirrCalculationResult.kt
@@ -1,0 +1,8 @@
+package ee.tenman.portfolio.service
+
+data class XirrCalculationResult(
+  val processedDates: Int,
+  val processedInstruments: Int,
+  val failedCalculations: List<String> = emptyList(),
+  val duration: Long,
+)
